requirements.txt
pygame==2.5.2
torchrl-nightly==2024.4.24
torch==2.2.2
tensorflow==2.5.0
gym==0.26.2
tensordict==0.3.2
